requirements.txt
aiogram==3.0.0
beautifulsoup4==4.12.3
requests==2.32.2
langchain==0.2.1
openai==1.30.1
faiss-cpu==1.8.0
python-dotenv==1.0.1